human always turns right
more predictability to the way the human moves makes the game a lot more
fun. I moved the gate over a few pixels so that the human wouldn't get
stuck in a infinite loop. also, being able to move while the human is
running helps to knock them out of infinite loops.